Fix type of allocated memory in harness for payloadSize

+++ libraries/standard/mqtt/cbmc/proofs/MQTT_GetSubAckStatusCodes/MQTT_GetSubAckStatusCodes_harness.c
@@ -37,7 +37,7 @@ void harness()
     __CPROVER_assume( isValidMqttPacketInfo( pSubackPacket ) );
 
     pPayloadStart = mallocCanFail( sizeof( uint8_t * ) );
-    pPayloadSize = mallocCanFail( sizeof( uint8_t ) );
+    pPayloadSize = mallocCanFail( sizeof( size_t ) );
 
     MQTT_GetSubAckStatusCodes( pSubackPacket,
                                pPayloadStart,
